% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r
\name{build_df}
\alias{build_df}
\title{Read *selected* tables from MMEX database and combine it as a tibble.}
\usage{
build_df(db_path)
}
\arguments{
\item{db_path}{path to MMEX database}
}
\value{
tibble
}
\description{
Note that this does not join all tables in the db.
It only joins
- checkingaccount
- accountlist
- category
- payee
Also this does *not* process for my preference, which will be done 'read_mmex_db'
}
