% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r
\name{get_category}
\alias{get_category}
\title{Extracts category}
\usage{
get_category(string_vec)
}
\arguments{
\item{string_vec}{string of categories}
}
\value{
string vector of category. Should not have any NA
}
\description{
Extracts category
}
