% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r
\name{get_subcategory}
\alias{get_subcategory}
\title{Extracts subcategory}
\usage{
get_subcategory(string_vec)
}
\arguments{
\item{string_vec}{string of categories}
}
\value{
string vector of subcategory. Returns NA_character_ if missing
}
\description{
Extracts subcategory
}
