% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r
\name{is_category_valid}
\alias{is_category_valid}
\title{Check whether the format is valid with only 0 or 1 ':'}
\usage{
is_category_valid(string_vec)
}
\arguments{
\item{string_vec}{string of categories}
}
\value{
logical
}
\description{
Should be either 'category' or 'category:subcategory'
}
