% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r
\name{list_table}
\alias{list_table}
\title{List all tables in mmex db.}
\usage{
list_table(db_path)
}
\arguments{
\item{db_path}{db file path}
}
\value{
character vector
}
\description{
List all tables in mmex db.
}
