% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r
\name{query_db}
\alias{query_db}
\title{Send query to mmex database and get the result}
\usage{
query_db(sql_query, db_path)
}
\arguments{
\item{sql_query}{SQL query}

\item{db_path}{db file path}
}
\value{
tibble from SQL query
}
\description{
Send query to mmex database and get the result
}
