% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r
\name{read_category_from}
\alias{read_category_from}
\title{Read category as a df}
\usage{
read_category_from(db_path)
}
\arguments{
\item{db_path}{path to mmex db file}
}
\value{
tibble with category information
}
\description{
This will run recursive query to get multi-level categories
}
