% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r
\name{read_mmex_db}
\alias{read_mmex_db}
\title{Read selected tables from MMEX database and combine it as a tibble.}
\usage{
read_mmex_db(db_path)
}
\arguments{
\item{db_path}{path to MMEX database}
}
\value{
tibble
}
\description{
Note that this does not join all tables in the db.
It only joins
- checkingaccount
- accountlist
- category
- payee
}
