% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r
\name{read_table_from}
\alias{read_table_from}
\title{Get a specific table from MMEX db using `query_db()`}
\usage{
read_table_from(tbl_name, db_path)
}
\arguments{
\item{tbl_name}{table within mmex db}

\item{db_path}{path to mmex db file}
}
\value{
tibble
}
\description{
To find the list of tables, use `list_table(db_path)`
The name assumes that pipe is used 'table_name' |> read_table_from(db_path)
}
